fix: show exact prices without rounding (e.g. 39.90 not 40)
_parse_price now returns float for non-integer prices.
plans.price column migrated from INTEGER to REAL.
UI formats price with up to 2 decimal places when needed.

+++ db.py
@@ -20,7 +20,7 @@ def init_db(db_path=None):
                 id         INTEGER PRIMARY KEY,
                 carrier    TEXT NOT NULL,
                 plan_name  TEXT NOT NULL,
-                price      INTEGER,
+                price      REAL,
                 data_gb    INTEGER,
                 minutes    TEXT,
                 extras     TEXT,
